+++ src/testcurl/test_digestauth_sha256.c
@@ -194,9 +194,9 @@ testDigestAuth (void)
   }
   while (off < 8)
   {
-    len = read (fd,
-                rnd + off,
-                8 - off);
+    len = (size_t) read (fd,
+                         rnd + off,
+                         8 - off);
     if (len == (size_t) -1)
     {
       fprintf (stderr,
@@ -262,8 +262,8 @@ testDigestAuth (void)
   }
   snprintf (url,
             sizeof (url),
-            "http://127.0.0.1:%d/bar%%20foo?key=value",
-            (int) port);
+            "http://127.0.0.1:%u/bar%%20foo?key=value",
+            (unsigned int) port);
   c = curl_easy_init ();
   curl_easy_setopt (c, CURLOPT_URL, url);
   curl_easy_setopt (c, CURLOPT_WRITEFUNCTION, &copyBuffer);
